+++ components/Templates/TimelineTemplate/style.tsx
@@ -1,0 +1,13 @@
+import styled from '@emotion/styled';
+
+export const TimelineContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  padding-bottom: 60px;
+`;
+
+export const TimelineTitleArea = styled.div`
+  display: flex;
+  justify-content: space-between;
+  padding: 36px 12px 0px 12px;
+`;
